# Fix Issue 19405 - Speed up backwards UTF-8 decoding in stripRight & make nogc nothrow for strings