Benchmark uses same test strings as https://github.com/dlang/phobos/pull/6770. ASCII timings were unaffected.

**string benchmark, ldc2:**

| compiler       | function       | data        | time                                |
|----------------|----------------|-------------|-------------------------------------|
| ldc2 -O2       | old_stripRight | non_ascii   | 1 sec, 362 ms, 31 μs, and 4 hnsecs  |
| ldc2 -O2       | new_stripRight | non_ascii   | 659 ms, 959 μs, and 4 hnsecs        |

**string benchmark, dmd:**

| compiler       | function       | data        | time                                |
|----------------|----------------|-------------|-------------------------------------|
| dmd -O -inline | old_stripRight | non_ascii   | 1 sec, 728 ms, and 184 μs           |
| dmd -O -inline | new_stripRight | non_ascii   | 859 ms, 895 μs, and 5 hnsecs        |